backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.1.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0

# Testing and Development
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0
coverage==7.3.2

# API Documentation
drf-spectacular==0.27.0

# Development Tools
django-debug-toolbar==4.2.0 